manager: Use Long for versionCode to support values exceeding Int.MAX
Fixed an issue where versionCode would get stuck at 2147483647. Updated data type to Long to support larger version numbers.

+++ manager/app/src/main/java/me/weishu/kernelsu/data/model/RepoModule.kt
@@ -28,6 +28,6 @@ data class RepoModule(
     val createdAt: String,
     val latestRelease: String,
     val latestReleaseTime: String,
-    val latestVersionCode: Int,
+    val latestVersionCode: Long,
     val latestAsset: ReleaseAsset?,
 )
