+++ manager/app/src/main/java/me/weishu/kernelsu/ui/util/module/LatestVersionInfo.kt
@@ -1,7 +1,7 @@
 package me.weishu.kernelsu.ui.util.module
 
 data class LatestVersionInfo(
-    val versionCode: Int = 0,
+    val versionCode: Long = 0L,
     val downloadUrl: String = "",
     val changelog: String = ""
 )
